AlertView#showLowBatteryAlert param type changed to Int

+++ app/src/main/java/com/j380/alarm/view/AlertView.kt
@@ -6,5 +6,5 @@ interface AlertView {
 
     fun hideAlert()
 
-    fun showLowBatteryAlert(batteryLevel: Float)
+    fun showLowBatteryAlert(batteryLevel: Int)
 }
